AG-9558 - Review fixes
