+++ packages/ag-charts-community/src/chart/axis/axis.ts
@@ -45,7 +45,8 @@ import { ChartAxisDirection } from '../chartAxisDirection';
 import { CartesianCrossLine } from '../crossline/cartesianCrossLine';
 import type { CrossLine } from '../crossline/crossLine';
 import type { AnimationManager } from '../interaction/animationManager';
-import { type PointerInteractionEvent, REGIONS } from '../interaction/interactionManager';
+import { REGIONS } from '../interaction/chartRegions';
+import { type PointerInteractionEvent } from '../interaction/interactionManager';
 import { calculateLabelBBox, calculateLabelRotation, getLabelSpacing, getTextAlign, getTextBaseline } from '../label';
 import { Layers } from '../layers';
 import type { AxisLayout } from '../layout/layoutService';
