+++ packages/ag-charts-community/src/chart/interaction/chartRegions.ts
@@ -1,0 +1,15 @@
+export const REGIONS = {
+    TITLE: 'title',
+    SUBTITLE: 'subtitle',
+    FOOTNOTE: 'footnote',
+    LEGEND: 'legend',
+    NAVIGATOR: 'navigator',
+    PAGINATION: 'pagination',
+    ROOT: 'root',
+    SERIES: 'series',
+    HORIZONTAL_AXES: 'horizontal-axes',
+    VERTICAL_AXES: 'vertical-axes',
+    TOOLBAR: 'toolbar',
+} as const;
+
+export type RegionName = (typeof REGIONS)[keyof typeof REGIONS];
