+++ packages/ag-charts-community/src/module-support.ts
@@ -39,6 +39,7 @@ export * from './chart/data/processors';
 export * from './chart/data/aggregateFunctions';
 export * from './chart/updateService';
 export * from './chart/layout/layoutService';
+export * from './chart/interaction/chartRegions';
 export * from './chart/interaction/animationManager';
 export * from './chart/interaction/chartEventManager';
 export * from './chart/interaction/contextMenuRegistry';
